# server/public/openApi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/system/dept": {
      "post": {
        "operationId": "DeptController_create",
        "summary": "部门管理-创建",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeptDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "部门管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "put": {
        "operationId": "DeptController_update",
        "summary": "部门管理-更新",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDeptDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "部门管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/dept/list": {
      "get": {
        "operationId": "DeptController_findAll",
        "summary": "部门管理-列表",
        "parameters": [
          {
            "name": "deptName",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "部门管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/dept/{id}": {
      "get": {
        "operationId": "DeptController_findOne",
        "summary": "部门管理-详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "部门管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "delete": {
        "operationId": "DeptController_remove",
        "summary": "部门管理-删除",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "部门管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/dept/list/exclude/{id}": {
      "get": {
        "operationId": "DeptController_findListExclude",
        "summary": "部门管理-黑名单",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "部门管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
...
    "/system/menu": {
      "post": {
        "operationId": "MenuController_create",
        "summary": "菜单管理-创建",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMenuDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "put": {
        "operationId": "MenuController_update",
        "summary": "菜单管理-修改",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/menu/list": {
      "get": {
        "operationId": "MenuController_findAll",
        "summary": "菜单管理-列表",
        "parameters": [
          {
            "name": "menuName",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/menu/treeselect": {
      "get": {
        "operationId": "MenuController_treeSelect",
        "summary": "菜单管理-树表",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/menu/roleMenuTreeselect/{menuId}": {
      "get": {
        "operationId": "MenuController_roleMenuTreeselect",
        "summary": "菜单管理-角色-树表",
        "parameters": [
          {
            "name": "menuId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/menu/{menuId}": {
      "get": {
        "operationId": "MenuController_findOne",
        "summary": "菜单管理-详情",
        "parameters": [
          {
            "name": "menuId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "delete": {
        "operationId": "MenuController_remove",
        "summary": "菜单管理-删除",
        "parameters": [
          {
            "name": "menuId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
...
    "/system/post": {
      "post": {
        "operationId": "PostController_create",
        "summary": "岗位管理-创建",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "岗位管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "put": {
        "operationId": "PostController_update",
        "summary": "岗位管理-更新",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "岗位管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/post/list": {
      "get": {
        "operationId": "PostController_findAll",
        "summary": "岗位管理-列表",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "当前分页",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "orderByColumn",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAsc",
            "required": false,
            "in": "query",
            "description": "排序规则",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListPostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "岗位管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/post/{id}": {
      "get": {
        "operationId": "PostController_findOne",
        "summary": "岗位管理-详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "岗位管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/post/{ids}": {
      "delete": {
        "operationId": "PostController_remove",
        "summary": "岗位管理-删除",
        "parameters": [
          {
            "name": "ids",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "岗位管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/post/export": {
      "post": {
        "operationId": "PostController_export",
        "summary": "导出岗位管理xlsx文件",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListPostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "岗位管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Nest-Admin",
    "description": "Nest-Admin 接口文档",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "请在请求头中携带 JWT 令牌，格式：Bearer <token>"
      }
    },
    "schemas": {
...
      "DateParamsDTO": {
        "type": "object",
        "properties": {}
      },
...
      "CreateDeptDto": {
        "type": "object",
        "properties": {
          "parentId": {
            "type": "number"
          },
          "deptName": {
            "type": "string"
          },
          "orderNum": {
            "type": "number"
          },
          "leader": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "parentId",
          "deptName",
          "orderNum"
        ]
      },
      "UpdateDeptDto": {
        "type": "object",
        "properties": {
          "parentId": {
            "type": "number"
          },
          "deptName": {
            "type": "string"
          },
          "orderNum": {
            "type": "number"
          },
          "leader": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "deptId": {
            "type": "number"
          }
        },
        "required": [
          "parentId",
          "deptName",
          "orderNum"
        ]
      },
...
      "CreateMenuDto": {
        "type": "object",
        "properties": {
          "menuName": {
            "type": "string"
          },
          "orderNum": {
            "type": "number"
          },
          "parentId": {
            "type": "number"
          },
          "path": {
            "type": "string"
          },
          "query": {
            "type": "string"
          },
          "component": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "menuType": {
            "type": "string"
          },
          "isCache": {
            "type": "string"
          },
          "isFrame": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "visible": {
            "type": "string"
          },
          "perms": {
            "type": "string"
          }
        },
        "required": [
          "menuName",
          "parentId",
          "isFrame"
        ]
      },
      "UpdateMenuDto": {
        "type": "object",
        "properties": {
          "menuName": {
            "type": "string"
          },
          "orderNum": {
            "type": "number"
          },
          "parentId": {
            "type": "number"
          },
          "path": {
            "type": "string"
          },
          "query": {
            "type": "string"
          },
          "component": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "menuType": {
            "type": "string"
          },
          "isCache": {
            "type": "string"
          },
          "isFrame": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "visible": {
            "type": "string"
          },
          "perms": {
            "type": "string"
          },
          "menuId": {
            "type": "number"
          }
        },
        "required": [
          "menuName",
          "parentId",
          "isFrame",
          "menuId"
        ]
      },
...
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "remark": {
            "type": "string"
          },
          "postSort": {
            "type": "number"
          }
        },
        "required": [
          "postSort"
        ]
      },
      "ListPostDto": {
        "type": "object",
        "properties": {
          "pageNum": {
            "type": "number",
            "description": "当前分页",
            "default": 1
          },
          "pageSize": {
            "type": "number",
            "description": "每页数量",
            "default": 10
          },
          "params": {
            "description": "时间范围",
            "allOf": [
              {
                "$ref": "#/components/schemas/DateParamsDTO"
              }
            ]
          },
          "orderByColumn": {
            "type": "string",
            "description": "排序字段"
          },
          "isAsc": {
            "type": "string",
            "description": "排序规则"
          }
        },
        "required": [
          "pageNum",
          "pageSize"
        ]
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "remark": {
            "type": "string"
          },
          "postSort": {
            "type": "number"
          },
          "postId": {
            "type": "number"
          }
        },
        "required": [
          "postSort",
          "postId"
        ]
      },
...
    }
  }
}